LPS-28999 Keep the organization's logo, as visually it is attached to the organisation itself, not the sites

+++ portal-impl/src/com/liferay/portal/service/impl/LayoutSetLocalServiceImpl.java
@@ -148,7 +148,11 @@ public class LayoutSetLocalServiceImpl extends LayoutSetLocalServiceBaseImpl {
 
 		// Logo
 
-		imageLocalService.deleteImage(layoutSet.getLogoId());
+		if (group.isStagingGroup() || !group.isOrganization() ||
+			!group.isSite()) {
+
+			imageLocalService.deleteImage(layoutSet.getLogoId());
+		}
 
 		// Layout set
 
@@ -157,7 +161,12 @@ public class LayoutSetLocalServiceImpl extends LayoutSetLocalServiceBaseImpl {
 		if (!group.isStagingGroup() && group.isOrganization() &&
 			group.isSite()) {
 
-			addLayoutSet(group.getGroupId(), privateLayout);
+			LayoutSet newLayoutSet = addLayoutSet(
+				group.getGroupId(), privateLayout);
+
+			newLayoutSet.setLogoId(layoutSet.getLogoId());
+
+			layoutSetPersistence.update(newLayoutSet, false);
 		}
 
 		// Counter
